requirements.txt
python-dotenv==1.0.0
requests==2.31.0
supabase==2.0.3
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio==3.4.3
tenacity==8.2.3
